--- tests/test_flunt.py
@@ -1,5 +1,0 @@
-from flunt import __version__
-
-
-def test_version():
-    assert __version__ == "0.2.0"
